+++ benchmark/benches/bench_query_posvel_soa_unpack.jl
@@ -1,8 +1,8 @@
 
 function setup_query_posvel_soa_unpack(n_entities::Int)
     world = World(
-        Position => Storage{StructArray},
-        Velocity => Storage{StructArray};
+        Position => Storage(StructArray),
+        Velocity => Storage(StructArray);
         boxed=BOXED,
     )
     for i in 1:n_entities