+++ ext/ArkCUDAInterop.jl
@@ -11,4 +11,23 @@ function Ark._gpuvector_hostwrap(mem::CuVector{T,CUDA.UnifiedMemory}) where {T}
     return unsafe_wrap(Vector{T}, mem)
 end
 
+function Ark._gpuvector_pinned_device(::Val{:CUDA}, ordinal::Integer)
+    return CuDevice(ordinal)
+end
+
+function Ark._gpuvector_ordinal(dev::CuDevice)
+    return CUDA.deviceid(dev)
+end
+
+function Ark._gpuvector_withdev(f, dev::CuDevice)
+    old = CUDA.device()
+    old == dev && return f()
+    CUDA.device!(dev)
+    try
+        return f()
+    finally
+        CUDA.device!(old)
+    end
+end
+
 end