+++ test/setup_large.jl
@@ -24,7 +24,7 @@ const DEFAULT_WORLD_BOXED = Ref(first(WORLD_MODES))
 struct FakeComp{N} end
 const N_fake = 300
 const fake_types = [FakeComp{i} for i in 1:N_fake]
-const fake_storage = [Storage{WrappedVector} for i in 1:N_fake]
+const fake_storage = [Storage(WrappedVector) for i in 1:N_fake]
 const M_mask = ceil(Int, N_fake / 64)
 const offset_ID = (M_mask - 1) * 64 - 1
 
@@ -36,24 +36,24 @@ function TestWorld(
 )
     raw_types = map(arg -> arg isa Type ? arg : arg.first, comp_types)
     types = map(Ark._unwrap_relation_type, raw_types)
-    storages = map(arg -> arg isa Type ? Storage{WrappedVector} : arg.second, comp_types)
+    storages = map(arg -> arg isa Type ? Storage(WrappedVector) : arg.second, comp_types)
     relation_types = map(Ark._unwrap_relation_type, filter(Ark._declares_relation, raw_types))
     storages = collect(Any, storages)
     for i in 1:length(storages)
-        if isbitstype(types[i]) && storages[i] == Storage{WrappedVector}
-            storages[i] = Storage{GPUVector{:CPU}}
+        if isbitstype(types[i]) && storages[i] == Storage(WrappedVector)
+            storages[i] = Storage(GPUVector{:CPU})
             break
         end
     end
     for i in 1:length(storages)
-        if storages[i] == Storage{WrappedVector} && isbitstype(types[i]) && fieldcount(types[i]) > 0
-            storages[i] = Storage{DiskVector}
+        if storages[i] == Storage(WrappedVector) && isbitstype(types[i]) && fieldcount(types[i]) > 0
+            storages[i] = Storage(DiskVector)
             break
         end
     end
     for i in 1:length(storages)
-        if storages[i] == Storage{StructArray}
-            storages[i] = Storage{GPUStructArray{:CPU}}
+        if storages[i] == Storage(StructArray)
+            storages[i] = Storage(GPUStructArray{:CPU})
         end
     end
     storages = Tuple(storages)